+++ crates/compiler/tests/mml/subroutines.rs
@@ -1254,3 +1254,62 @@ A @d o6c !s o4c o6d o4e
         BytecodeError::NoteOutOfRange(note("d6"), note("c4")..=note("b4")).into(),
     );
 }
+
+#[test]
+fn song_loop_point_in_nested_subroutine_panic_bugfix() {
+    // found using rust-fuzz
+    assert_one_subroutine_error_in_mml(
+        r##"
+@0 dummy_instrument
+
+!s1 @0 r L r
+!s2 !s1 r
+
+A !s2
+"##,
+        "!s1",
+        10,
+        ChannelError::CannotSetLoopPoint,
+    );
+}
+
+#[test]
+fn tail_call_subroutine_with_compile_error_panic_bugfix() {
+    let dummy_data = dummy_data();
+
+    // found using rust-fuzz
+    let r = compiler::songs::compile_mml_song(
+        r##"
+@0 dummy_instrument
+
+!s1 !s2
+!s2 {cP} @0
+
+A !s2
+"##,
+        "",
+        None,
+        &dummy_data.instruments_and_samples,
+        &dummy_data.pitch_table,
+    );
+    assert!(r.is_err());
+
+    let r = compiler::songs::compile_mml_song(
+        r##"
+@0 dummy_instrument
+
+!s1 !s2
+!s2 !s3
+!s3 !s4
+!s4 !error
+!error {cP} @0
+
+A !s1
+"##,
+        "",
+        None,
+        &dummy_data.instruments_and_samples,
+        &dummy_data.pitch_table,
+    );
+    assert!(r.is_err());
+}